removed the need for maintaining a list of plugins
The plugin directory is now set in the config file
and all .py file in that directory will be loaded.
Plugins can be updated, added, or removed while
the process is running. The plugins are now
controlled entirely by their existence in the
plugins directory. A plugin author only needs to
edit the plugin itself without having to remember
(or be allowed) to edit the __init__.py file.